requirements.txt
# FastAPI and ASGI server
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
motor==3.3.2
pymongo==4.6.1
dnspython==2.5.0  # Required for MongoDB Atlas (mongodb+srv://)

# Authentication (JWT for sessions, no passwords needed with Google OAuth)
python-jose[cryptography]==3.3.0

# Environment & Config
python-dotenv==1.0.0

# Networking & Async HTTP (for Mistral/Qwen API calls)
httpx==0.26.0
aiohttp==3.9.1

# Existing dependencies
networkx==3.2.1
matplotlib==3.8.2
google-auth==2.25.2
pytest==7.4.3
pydantic==2.5.3

# Optional: For better performance in production
# gunicorn==21.2.0
# python-multipart==0.0.6  # For file uploads if needed
